requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
splitwise==2.4.0
crewai==0.11.0
python-multipart==0.0.6
pillow==10.1.0
transformers==4.36.0
# torch==2.1.1
pydantic==2.5.2
python-jose[cryptography]==3.3.0
